base: Split out the store module into smaller submodules
